ggsn-test: Use ggsn local ip as announced DNS
This way the address is reachable for ggsn system and can route the
packet correctly.

This commit fixes runtime failures of
GGSN_Tests.TC_pdp4_act_deact_gtpu_access when running inside the
container, since that test uses the DNS address to attempt a ping and
waits for a response. If the address is non reachable/routable from
osmo-ggsn tun, then the packet is dropped by the kernel and no ping
response is sent back.