ttcn3-msc-test: Disable MM INFO
In the past, MM INFO transmission was broken, so the tests were
written without expecting it.  Now, the bug in osmo-msc has been
fixed.

With the config file enabling it, and the tests not expecting MM INFO,
a lot of them failed. Let's disable MM INFO for now until the tests
can deal with it