Add Dockerfile for gerrit and monkey patch
The openid support doesn't allow to add custom providers but we like the
convenience of a single click. Take the official docker image and patch
the HTML template. It seems to have changed little over the time, and it
might continue like that...

I have filed https://bugs.chromium.org/p/gerrit/issues/detail?id=8828 so
that we can drop this hack in the future...

Change-Id: I713948fbb93355c2e33e3b92969e2389cb88c938