Introduce Dockerfile for the GR-GSM project
This change introduces a new sub-project - 'gr-gsm-master'.
It was decided to use Arch Linux as the base image, because it
provides the recent version of both UHD and GNU Radio without
tons of dependencies, such as Xorg.

Note: as there in no libosmocore package in Arch Linux, it will
be built from the latest source code.

Change-Id: I115b59dc9eaec5d132f3f64c4ee321173459e9ed