ttcn3-sip-test/jenkins.sh: remove invalid -M arg
osmo-sip-connector does not have a -M argument, it was probably
confused with osmo-msc's -M, --mncc-sock-path argument.

osmo-sip-connector printed out the following when started with -M, but
otherwise functioned normally. Because it was started with Docker in
background, the message did not get written to the build log.

osmo-sip-connector: invalid option -- 'M'

Change-Id: Iee983a8aafe8cade67b76927721fc26e5af966d7