debian-*-build: don't download "Release" file
Stop jenkins jobs from failing, due to the download errors we are
getting from time to time, by simply not downloading these files. They
are redundant, because we are already running 'docker build --pull'
before trying to build the debian-*-build images [1], and the docker
images are released at least as often, as the "Release" file gets
updated [2].

This is in line with how it is done in osmo-ci.git.

[1]: Change-Id: I1076bbb7d77bdc99f5d60d641c09ce965fb9dfc6
[2]: https://hub.docker.com/_/debian

Change-Id: Ieedb310814412f235e56bdc65517404842eac00f