repo-install-test: disable osmo-ggsn service
OsmoGGSN is not able to use the tun4 device from the default config in
docker. Since the more strict config checking in [1], it does not just
report a warning, but fails to start:

	<0002> ggsn.c:189 APN(internet): Opening TUN device tun4
	<0001> tun.c:184 errno=2/No such file or directory open() failed
	<0002> ggsn.c:191 APN(internet): Failed to configure tun device
	Error occurred during reading the below line:
	  no shutdown
	Failed to open config file: '/etc/osmocom/osmo-ggsn.cfg'

Fix the repo install test jenkins job by not checking osmo-ggsn anymore.
In theory, we could probably create the tun device on the host, and
mount it inside the docker container. But that would require some
additional logic to clean it up properly, and it does not seem worth the
effort right now.

[1] libosmocore Ic9c1b566ec4a459f03e6319cf369691903cf9d00
Change-Id: I15c7a2f38a21cd96c9852ad26f4b4a81798e6593