remsim: Ensure non-overlapping junit-xml filenames
The default filename pattern of the JUnit log plugins of TTCN-3 is
"junit-xml-%u.xml" where %u represents the PID of the process.  This
is typically a good idea as the PID doesn't wrap that quickly.  However,
if you're executing a new docker container for every test suite run,
then chances are high that two subsequent runs of the test suite
container might end up with the same PID and hence the smae filename.

This results in the output file being overwritten, which is what's
responsible for many missing  RemsimServer_Tests results on the
jenkins.osmocom.org test results analyzer.

Change-Id: I9b6d1b5a943eb4e058ad2bf7cf8acee7b70e68b6
Closes: OS#4205