ttcn3-bts-test/oml: fix osmo-bts.cfg: osmo-bts-trx has no DSP/HW MS Power Control
It was noticed that most of the OML related TTCN-3 test cases do not
pass anymore, while some of them are green for the release builds.

The culprit is I49706926b1e962b18791174627bc3cc0cd0cd9d5, which did
not ensure backwards compatibility with older configuration files.
Let's remove 'ms-power-control dsp' and use the defaults.

Change-Id: I940df510d803d0eaa063dd5145c2f9a4c5d0543f