migrate ttcn3-msc-test from 172.18.1.0/24 to 172.18.20.0/24
When introducing IPv6 support, we map the third digit of the IPv4
address (X) to the 6th byte of an IPv6 prefix "fd02:db8:X::/64"

However, the docker daemon seems to use "fd02:db8:1::/64" internally
for its default network, so creating a docker network with the same
IP address is failing.

Let's move the MSC test suite to another sub-net (1->20) to avoid
related problems.

Change-Id: I9c5f9b96d5523eae09f3f2e6c813e9e0d047f9ab