ttcn3-ggsn-test: Use IPv6 address of the tester as "DNS1"
We cannot rely on every test slave always having proper IPv6
conenctivity to a public DNS server, and that DNS server to respond
to ICMPv6 echo request.  We're testing half of the internet instead
of our GGSN.  Let's use the IPv6 address of the docker container
running the test sute, which is guaranteed to always exist.  We have
no problems with routing, as the ICMPv6 is hand-crafted and encapsulated
into GTP before it's sent to the kernel IP stack - so basically on the
transmit side, the kernel doesn't know anything about the destination
address of the encapsulated IPv6 packet.

Closes: OS#4713
Change-Id: I1377ffa2cc6980168297d5ff5c81fa5dad94d7a9