osmo-*-master: Invalidate docker cache when CentOS repo changes
We currently only invalidate the cache if the Debian repository changes,
but not if the CentOS repository changes.  This leads to strange
artefacts, such as CentOS containers for osmo-bts-master using an
old libosmocore, despite a new one being available in the repo.

As Dockerfiles don't support conditional "ADD", we need to actually
add both the CentOS and the Debian repsotiry file, so any change
in either of the two will now (unneccessarily) invalidate the cache
for both Debian and CentOS builds.

Change-Id: I7b13644afb9d3be3b94184193177e003c13718d3