ttcn3-bts-test: fix: sleep 1 second before respawning the IUT
There has been a lot of changes merged to both osmo-{bsc,bts}, so
now the later seems to be respawning faster then before.  Because
of that, osmo-bts-trx is now reaching the limit of 500 runs on
Jenkins right in the middle of ttcn3-bts-test execution:

  respawn.sh: exiting after 500 runs

causing the remaining test cases to fail due to:

  "BTS_Tests.ttcn:253 : Timeout waiting for RSL bring up"

Most likely, the IUT respawns several times in the windows between
the test cases, i.e. after TITAN finishes one test case and before
it starts another.  Let's instruct respawn.sh to sleep 1 second
before starting the process again.

Change-Id: I09e5a8e7c4d28424cd0d004a8bae3cf0ade2d838