osmo-gsm-tester: open5gs deps only for x86_64
mongo-db is only available for x86_64 from their third party
repository. Don't attempt to install it for another architecture. As
this is part of the open5gs dependencies, don't install any of them
either.

This should fix the currently failing "update-osmo-ci-on-slaves"
jenkins job. I've considered disabling the build of the osmo-gsm-tester
container for ARM altogether, but the osmo-gsm-tester manual explicitly
mentions ARM trails.

Change-Id: Ieed36b1adf4b9ae147c7ee7ade22d4855c3dbac8