Don't run "dpkg --add-architecture i386" on ARM
Building debian-stretch-jenkins on ARM currently results in a few
libraries getting installed for the i386 architecture. This results in
the following build failures for master-osmo-trx and master-osmo-pcu:

  checking for LIBGNUTLS... no
  configure: error: Package requirements (gnutls >= 2.12.0) were not met:

  No package 'gnutls' found

It is weird that we are only seeing this now, the dpkg-line has been in
debian-stretch-jenkins since the Dockerfile was added to osmo-ci four
years ago in Icb0406f96b0c18e77be51ad8317c2668fb23a45e. I'm also
wondering why it was added in the first place, and if we need it
anymore. In any case, running this on anything but x86_64 does not make
sense, so disable it there.

Change-Id: I8d890b7ebc7b99a4c79779709d1e56e0a5a6d4b9