debian-stretch-build-dist: osmo-gsm-manuals deps
Add dependencies for osmo-gsm-manuals, so we can build the release
tarballs for it (autoreconf -fi; ./configure; make dist-bzip2) in the
related jenkins job. During update of the list of projects for which we
build these tarballs, it became apparent that it was missing. Following
patches will add dependencies needed to build release tarballs for all
other projects that were missing from the list.

Related: OS#5347
Change-Id: Iba2e71b2e757bc527561d0f3e4a1af5f024a3cd7