osmo-pcap-master: doesn't need libosmo-netif
Don't attempt to install libosmo-netif in this Dockerfile. I've noticed
this because we want to build osmo-pcap for CentOS7 with this container
to run the TTCN-3 tests, and we don't have libosmo-netif for CentOS7.

Add pkgconfig(libosmogb), which is required to build osmo-pcap
according to configure.ac. In Debian, this gets installed as part of
libosmocore-dev.

Related: SYS#5754
Change-Id: I9c3a3b43ee7c25c06042f3303b9edb4005e7db31