osmocom-bb-host-master: modernize
Do not set up the Osmocom repository inside osmocom-bb-host-master, it
is done in debian-bullseye-build. Without this patch, we actually
installed the debian 9 version of libosmocore-dev in the debian 11
container.

This is part 1/2 of fixing TTCN3-centos-bts-test, which has been failing
since OSMOCOM_REPO_MIRROR was changed in osmo-ci.git. Part 2 is in
osmo-ci I8201f34a0f9747eea78c358a626cc3981d7407c4.

Change-Id: I78e0445a1b82e5cd72ecaa86e8fe25516b19f485