osmo-gsm-tester: use debian's soapysdr-module-lms7
After switching from build.opensuse.org to obs.osmocom.org's published
packages, the soapysdr0.7-module-lms7 binary package we had does not
exist anymore.

I'm not sure where it came from, from the name it looks like it was
built from version 0.7 sources, but we only have a soapysdr_0.5.4
package in OBS that was added for debian 8/ubuntu 16.04 compatibility 4
years ago, and builds have been disabled for all architectures:
https://build.opensuse.org/package/show/network:osmocom:nightly/soapysdr

Instead of relying on this leftover binary package, use the one provided
by debian instead. This osmo-gsm-tester docker container is based on
debian-buster, which has soapysdr 0.6 packaged. If we should really need
0.7, we could switch it to bullseye or build it from source while
building the container, or add it back to obs.osmocom.org.

I verified that the container builds with this change.

Fix for:
  E: Unable to locate package soapysdr0.7-module-lms7

Change-Id: Ie2408258eda407d9db6c2f0f5fbe475861756778