release-tarball-build-dist: upgrade from debian 9
Let it use debian bullseye and rename the directory once, so we don't
need to rename it in the future when switching the debian version
again.

The reason I'm doing this now is that building with debian 9 failed
due to obs.osmocom.org related changes. It would be trivial to fix that,
but I decided to rather upgrade to the latest debian version.

Change-Id: Ib6039724a7200abce5252f95407047296b72cd51