release-tarball-build-dist: move osmo-ci code here
Add a jenkins.sh that does the same as the previous lengthy shell
command in the jenkins-job-builder file, and add
osmocom-release-tarballs.sh. This makes it easier to test the
script, and only patching one repository is needed from now on when
changing to a newer debian version.

Depends: osmo-ci I15f6c08b807f1462f91e95b624cdc1aad0592dd6
Change-Id: I4cf66097f508c029f688ec8af1cab2238d42f84f