fix_perms: use debian:bullseye
Use debian:bullseye instead of debian-bullseye-build. Building the
latter requires a lot of time, it's definitively not needed for just
running chmod. Use debian:bullseye, since we use it as base for most
images and it's therefore very likely that it's already downloaded
(and if not, it would get downloaded here).

Change-Id: I88f513a0a8ee04628842a7aa1b663c390d2032cd