debian-buster-jenkins: add pysim doc/lint depends
The lint and docs parts of pysim's contrib/jenkins.sh will run on a
different jenkins node than the tests (which has to run on simtester).
Add the depends for lint and docs to debian-buster-jenkins, so we can
run them in docker.

The python libraries construct, pytlv, pyyaml are needed so pylint
doesn't print false errors like "Undefined variable 'TLV'".

Related: OS#5497
Change-Id: Id5c75725d2fab46b29773fa4f637fa2d73fa7291