debian-buster-jenkins: add more pysim related deps
Install all depends to properly generate the manuals. Otherwise it
prints errors like the following and the resulting manual has empty
sections:
  /build/docs/shell.rst:329: ERROR: Failed to import "LinFixedEF" from "pySim.filesystem".
  No module named 'smartcard'

Get the requirements.txt directly from the git repository, so we don't
need to update the list of dependencies here. Remove the dependencies
from the list that I added recently for pysim, and which are are also
in requirements.txt.

Add swig, as otherwise pyscard fails to install.

Related: OS#5497
Change-Id: I13a0b0dc8be81bf4034b08b1ef1100f1420fa3a1