ttcn3-docker-run.sh: Use interface "up" flag, not operstate
Don't use the "operstate" sysfs attribute to determine if an interface
is "up", use the actual UP-flag (0x01) in the "flags" sysfs attribute.

The "operstate" attribute may at least occasionally be "unknown" and
remain in that state (causing jenkins jobs to wait indefinitely), while
the flags (which we don't look at before this patch) indicates it is "up".

This is a fixup to the below commit:

commit d2014603a736a17d6b92006b25aa41482365ef3d
Author: Harald Welte <laforge@osmocom.org>
Date:   Wed Feb 3 22:05:43 2021 +0100

    debian-stretch-titan: Wait for interface to be _up_ not just its existance

Change-Id: Ib5c3bbe470ce874217437c2518df5ae07f0d8301
Closes: OS#5803