debian-buster-jenkins-arm: new image
Add a minimal image to use on the raspberry pis, so it doesn't need to
spend 1-2 hours building debian-buster-jenkins on the raspberry pis,
full of things we don't actually use there.

I've verified that osmo-bts, osmo-pcu, osmo-trx build with this image
(the only projects which we build on the rpis).

Related: OS#5863
Change-Id: I4fb4b48b233acaef815c4c27ec6f17f12cfe836b