ttcn3-pgw-test: update open5gs-{nrf,smf,upf} config files
As stated in the ticket, both open5gs-{nrf,smf} daemons refuse to
start since build #326 because of the following error:

  [sbi] ERROR: TLS enabled but no server key (../lib/sbi/context.c:186)

The problem is that the recent open5gs (v2.6.0-39-g22be888 at the
moment of writing) is attempting to employ TLS verification for SBI
even if it's not explicitly configured in the config files.

  commit 05fbaf69587488e53b5e741a9ada9f9fa5749322
  Author: Sukchan Lee <acetcom@gmail.com>
  Date:   Sat Feb 18 10:58:29 2023 +0900

    [SBI] HTTP2-TLS verification - ConfFile Changed

Our config files are slightly outdated, so let's take a chance
to update them and specify the missing TLS params.  Check out
open5gs-{nrf,smf,upf}.yaml from the latest git version.

The updated files are confirmed to work with both:

* latest release for Debian v2.5.8, and
* latest git version v2.6.0-39-g22be888.

Change-Id: I27adbab6a6b95ddf4c6d39803e4f7dd079f11a4c
Related: OS#5913