debian-bullseye-jenkins: add iproute2
Add iproute2 to make "ss" available. It is used in the openbsc testsuite
to dump socket statistics. If it isn't available, it does not cause
tests to fail, but prints an error message that looks like it might be
the real cause for the test failures:

  > ss -tn
  /bin/sh: 1: ss: not found

I've also asked Harald if we should just retire the master-openbsc job,
but we want to keep it around to ensure new libosmocore builds with this
old codebase.

Change-Id: I306caeda9bcd4020734eb3b343420c8115958ded