osmo-uecups-master: make osmocom user avilable
Bring back the osmocom user that was removed earlier to fix a UID
conflict with the build user.

The osmocom user is being used after all, the TTCN-3 PGW tests use it to
start commands as "osmocom" user in osmo-uecups-master. This lead to e.g.
PGW_Tests.TC_createSession_ping4 failing.

Rename the user instead of adding a new one, so it can write log files
to the log dir as the same UID as the user that runs docker.

Fixes: 94f8a8 ("osmo-uecups-master: remove useradd")
Related: OS#6057
Change-Id: I2b29d7fedcf3e0a4c17bfbad16077098f94c529b