debian-*-erlang: Install libsctp to enable sctp support in erlang apps
The current "--enable-sctp" flag when building erlang consists of
supporting SCTP if the libsctp library is found at runtime, by
dynamically linking to it only when SCTP features are requested.

The erlang package in debian only lists libsctp as a Build-Depends, but
it is not installed by default, so apps wishing to use SCTP will fail:
"""
[error] CRASH REPORT Process <0.155.0> with 0 neighbours crashed with reason: no match of right hand value {error,{shutdown,{open,eprotonosupport}
"""

Installing the libsctp library fixes the issue.

Related: OS#5757
Change-Id: I5ca332f9064c16491116608577b3f68379d840e6