open5gs-master: Install open5gs-dbctl to /usr/bin
meson install doesn't install that file because it's aimed at
developers. However, debian package does install it manually in
debian/open5gs-common.install.
Hence, the app is available in open5gs-latest docker image, but not in
open5gs-master. Manually copy it to /usr/bin to have both behave the
same in this aspect.

Related: SYS#6588
Change-Id: I54a760d04e6522d843e6d65b94ef6067e502fbaa