ttcn3-pcu-test/sns: fix PCUIF version number (follow up patch)
Unfortunately we had to revert [1] because it worked for current master
but not for latest. The mistake here was to change the PCUIF version
number in PCU_Tests.cfg to PCUIF v12. This is indeed the correct version
for current master, but latest still uses v11. Also the change we made in
jenkins-sns only affected 2023q1 builds, it does not affect latest. This
is the reason why the previous patch broke latest.

This follow up patch now copies the approach we already successfully use
with the normal ttcn3-pcu-test / ttcn3-pcu-test-latest. (see also
jenkins.sh)

[1] I0b37f01f4c7bb829053231339e39ab734f4c8cbc

Change-Id: I1f94a0459e35d7b5632c81d7f7e2e60eb0d0229f
Related: OS#6275