ttcn3-s1gw-test: start osmo-s1gw container with --user=root
For some reason, on Arch with docker 26.1.3 it implicitly starts
with --user=root, while with docker 20.10.5 on Debian bookworm it
starts with --user=build.  This results in permission errors:

  + ip addr add 172.18.74.100/24 dev eth0
  RTNETLINK answers: Operation not permitted
  + true
  + ip addr add 172.18.74.200/24 dev eth0
  RTNETLINK answers: Operation not permitted

Pass --user=root explicitly to bring consistency and fix those.
Hopefully, ttcn3-s1gw-test will be passing on Jenkins with this patch.

Change-Id: I5cc67d1c3192fd55560c16c8cc44b13b2c9b3e98
Fixes: 9ca9bed "ttcn3-s1gw-test: fix using unassigned addr for osmo-s1gw"